+++ CHANGELOG.md
@@ -2,6 +2,7 @@ All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
+- Enforce non-zero price in ratechecker API.
 - Add new ratechecker API endpoint to check data load timestamp.
 
 ## 0.9.94 - 2017-02-02